Patch for #377; fixes a dangling-pointer bug in _ReleaseCelData
This is triggered by having one thread lock a graphics resource
(usually through a StarShipPtr) and then have another thread release
it.  The Frame data is gone, but pointers to that data (as well
as the Drawable itself, of course) remain.  If the 'locking' thread
then tries to use its Frame data, it will have end up accessing
garbage.  (The nulling of the DrawablePtr->Frame pointer doesn't help
because the Frame itself has been cached.)


git-svn-id: https://sc2.svn.sourceforge.net/svnroot/sc2/trunk/sc2@946 8092fc87-c524-0410-9efc-e669fe64eaf9