Fixed a harmless (as far as UQM is concerned) bug which caused unjustified warnings with .zip files generated from DOS/Windows.
Fixed a potentially memory-corrupting bug when in a .zip file a subdir
is described before a dir higher in the dir hierarchy.

Fixed a bug where the path to a file was only partially printed in
an fprintf used in error handling. (harmless)

Renamed some functions to better reflect their functionality.


git-svn-id: https://sc2.svn.sourceforge.net/svnroot/sc2/trunk/sc2@1197 8092fc87-c524-0410-9efc-e669fe64eaf9