Voiceove clip abstraction.
This is two simultaneous changes: CONVERSATION resources are made of 
three paths instead of one (phrases, clip directory, timestamps), and 
the synthetic, always-loaded '3dovoice' addon controls the use of them.

The intent is to eventually move uqm-0.x.0-voice.zip out of the core 
packages and into the addons with 3domusic.


git-svn-id: https://sc2.svn.sourceforge.net/svnroot/sc2/trunk/sc2@2978 8092fc87-c524-0410-9efc-e669fe64eaf9